requirements.txt
django==5.0.1
pytelegrambotapi==4.14.1
redis==5.0.1
requests==2.31.0
phonenumbers==8.13.27
django-phonenumber-field[phonenumbers]==7.3.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1